We are live!

+++ src/main.py
@@ -43,14 +43,18 @@ if __name__ == '__main__':
     log.info("Data has been updated since the last check, need to archive")
     archiver.archive(dir_to_backup)
     pcloud_client.upload(new_archive_name)
+
     if pcloud_client.is_file_identical(new_archive_name):
+        log.info("Backup success")
         storage.update_directory_hash(dir_to_backup, current_hash)
         storage.update_last_execution_date(date.get_today())
-        log.info("Backup success - Deleting the previous backup on pCloud")
-        pcloud_client.delete_file(previous_archive_name)
-    # else:
-    #     log.info("Backup failure - Deleting the corrupted file on pCloud")
-    #     pcloud_client.delete_file(new_archive_name)
+
+        if pcloud_client.is_file_present(previous_archive_name):
+            log.info("Deleting the previous backup on pCloud")
+            pcloud_client.delete_file(previous_archive_name)
+    else:
+        log.info("Backup failure - Deleting the corrupted file on pCloud")
+        pcloud_client.delete_file(new_archive_name)
 
     os.remove(new_archive_name)
